+++ extension.js
@@ -103,7 +103,7 @@ function activate( context )
             provider.add( rootFolder, match );
         } );
         status.hide();
-        provider.refresh();
+        provider.refresh( true );
     }
 
     function search( rootFolder, options, done )
@@ -242,6 +242,14 @@ function activate( context )
         iterateSearchList( searchList );
     }
 
+    function setButtons()
+    {
+        console.log( "setButtons" );
+        var expanded = vscode.workspace.getConfiguration( 'todo-tree' ).expanded;
+        vscode.commands.executeCommand( 'setContext', 'todo-tree-show-expand', !expanded );
+        vscode.commands.executeCommand( 'setContext', 'todo-tree-show-collapse', expanded );
+    }
+
     function showFlatView()
     {
         vscode.workspace.getConfiguration( 'todo-tree' ).update( 'flat', true, false ).then( function()
@@ -265,9 +273,30 @@ function activate( context )
         {
             return match.file !== filename;
         } );
+    }
 
         searchList = [ { file: filename } ];
         iterateSearchList();
+    }
+
+    function collapse()
+    {
+        vscode.workspace.getConfiguration( 'todo-tree' ).update( 'expanded', false, false ).then( function()
+        {
+            provider.clear();
+            addToTree( getRootFolder() );
+            setButtons();
+        } );
+    }
+
+    function expand()
+    {
+        vscode.workspace.getConfiguration( 'todo-tree' ).update( 'expanded', true, false ).then( function()
+        {
+            provider.clear();
+            addToTree( getRootFolder() );
+            setButtons();
+        } );
     }
 
     function register()
@@ -297,6 +326,8 @@ function activate( context )
         context.subscriptions.push( vscode.commands.registerCommand( 'todo-tree.refresh', rebuild ) );
         context.subscriptions.push( vscode.commands.registerCommand( 'todo-tree.showFlatView', showFlatView ) );
         context.subscriptions.push( vscode.commands.registerCommand( 'todo-tree.showTreeView', showTreeView ) );
+        context.subscriptions.push( vscode.commands.registerCommand( 'todo-tree.expand', expand ) );
+        context.subscriptions.push( vscode.commands.registerCommand( 'todo-tree.collapse', collapse ) );
 
         vscode.window.onDidChangeActiveTextEditor( function( e )
         {
@@ -342,6 +373,7 @@ function activate( context )
             {
                 provider.clear();
                 addToTree( getRootFolder() );
+                setButtons();
             }
         } ) );
 
@@ -350,6 +382,7 @@ function activate( context )
         var flat = vscode.workspace.getConfiguration( 'todo-tree' ).flat;
         vscode.commands.executeCommand( 'setContext', 'todo-tree-flat', flat );
 
+        setButtons();
         rebuild();
     }